Preliminary LEvents design
It might look kinda complicated at first so i've tried to explain every part of it in the method descriptions.

Still looking to simplify it down, and some of the features (like the event thread) will not be making it into version 1.

It's also really nice to have it templated, for type safety, but putting the whole thing in the header file is gonna be a pain. I will try and work around that as much as I can.