LEngine is static, but only in a private sense. It has a concept of the current engine being run, and this can be used in it's internal methods or in static accessor methods.
This way we shouldn't need to worry about crazyness in terms of static things getting messy, and it also lets other things access the needed stuff.

Maybe it would be best to have the Engine not static at all, and have the game just have a reference to it. That's a discussion for another time.